ADD : Ajout de blueprints pour flask
Dans ce commit, les api sont divisés en différents fichiers. Ceci est fait grace à l'ajout de blueprints. Un fichier requirements.txt a également été ajouté.
