+++ flask/api/hello_world.py
@@ -1,0 +1,10 @@
+from flask import Blueprint, request, jsonify, json
+from database import db
+
+
+app_hw = Blueprint('app_hw', __name__)
+
+
+@app_hw.route('/')
+def hello_world():
+    return 'Hello world!'
