+++ flask/app.py
@@ -1,93 +1,15 @@
 from flask import Flask, request
 from pymongo import MongoClient
-from bson import json_util
+from api import blueprints, database
 import json
-app=Flask(__name__)
-
-client = MongoClient("mongodb+srv://alexis:sixela@clusteralexis.7q0fj.mongodb.net/test")
-db = client.get_database('SmartFridge')
 
 
-@app.route('/')
-def hello_world():
-    return 'Hello world!'
+def create_app():
+    app = Flask(__name__)
 
+    for blueprint in blueprints:
+        app.register_blueprint(blueprint)
 
-@app.route('/client', methods=['GET', 'POST', 'PUT', 'DELETE'])
-def client():
-    records = db.Client
+    return app
 
-    if request.method == 'GET':
-        result = list(records.find())
-
-        for client in result:
-            client.pop('_id')
-        
-        return json.dumps(result)
-
-    elif request.method == 'POST':
-
-        parameters = request.json
-        print(parameters)
-
-        new_client = {
-        "name" : "test",
-        "firstName" : "test",
-        "email" : "test@gmail.com",
-        "pseudo" : "Miki",
-        "password" : "Miki1234",
-        "furniture" : {
-                        "Frigo" : {
-                            "temperature" : "4.3",
-                            "Hygrometry" : "68",
-                            "Products" : {
-                                    "Cordon bleu" : {
-                                        "ingredients" : ["jambon","fromage","chapelure"],
-                                        "expiration_date" : "20/12/21",
-                                        "weigth" : "300", 
-                                        "energy" : "250",
-                                        "fat" : "35",
-                                        "carbohydrates" : "10",
-                                        "proteins" : "29",
-                                        "minerals" : "2"                                         
-                                    },
-                                    "Pomme" : {
-                                        "ingredients" : [],
-                                        "expiration_date" : "25/12/21",
-                                        "weigth" : "300", 
-                                        "energy" : "250",
-                                        "fat" : "35",
-                                        "carbohydrates" : "10",
-                                        "proteins" : "29",
-                                        "minerals" : "2"                                         
-                                    }
-                            }
-                        }, 
-
-                        "Cabinet" : {
-                            "temperature" : "20",
-                            "Hygrometry" : "68",
-                            "Products" : {
-                                    "Cordon bleu" : {
-                                        "ingredients" : ["jambon","fromage","chapelure"],
-                                        "expiration_date" : "20/12/21",
-                                        "weigth" : "300", 
-                                        "energy" : "250",
-                                        "fat" : "35",
-                                        "carbohydrates" : "10",
-                                        "proteins" : "29",
-                                        "minerals" : "2"                                         
-                                    },
-                                    "Pomme" : {
-                                        "ingredients" : [],
-                                        "expiration_date" : "25/12/21",
-                                        "weigth" : "300", 
-                                        "energy" : "250",
-                                        "fat" : "35",
-                                        "carbohydrates" : "10",
-                                        "proteins" : "29",
-                                        "minerals" : "2"                                         
-            }}}}}
-            
-        records.insert_one(new_client)
-        return "client ajouté"
+app = create_app()
