+++ flask/requirements.txt
@@ -1,0 +1,10 @@
+bson==0.5.10
+click==8.0.3
+Flask==2.0.2
+itsdangerous==2.0.1
+Jinja2==3.0.2
+MarkupSafe==2.0.1
+pymongo==3.12.0
+python-dateutil==2.8.2
+six==1.16.0
+Werkzeug==2.0.2
